+++ src/app/types/index.ts
@@ -1,4 +1,3 @@
-export * from './data-table';
 export * from './context';
 export * from './ui';
 export * from './domain';
